server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
RPi.GPIO==0.7.1
python-dotenv==1.0.0
fake-rpi==0.7.1
psutil==5.9.5
Flask-JWT-Extended
Flask-SocketIO==5.3.4
python-socketio==5.8.0
requests
paho-mqtt
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
eventlet==0.33.3
paho-mqtt==1.6.1
Werkzeug==2.3.7
bcrypt==4.0.1
PyJWT==2.8.0